pyproject.toml
[project]
name = "ai-architect-prompt-engineering"
version = "0.1.0"
description = "ReAct/Self-Reflection agent with prompt debugging, optimization, and meta-prompting"
requires-python = ">=3.11,<3.12"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.11",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0.3",
    "tenacity>=8.5",
    "typer>=0.15.0",
    "rich>=13.9",
    "python-dotenv>=1.0",
]

[project.scripts]
prompt-engineering = "prompt_engineering.main:app"

[dependency-groups]
dev = [
    "black~=24.10.0",
    "isort>=6.0.1",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_engineering"]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.ruff]
target-version = "py311"
exclude = [".eggs", ".venv", "build", "dist", "outputs"]

[tool.ruff.lint]
select = ["E", "F", "W"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120
